Create TrueMockServer package-private class
Creates a single package-level class for all tests to use, rather than each test having their own private instance of the class.
